refactoring of main package name and *tAco class
