+++ HELP.md
@@ -1,7 +1,7 @@
 # Read Me First
 The following was discovered as part of building this project:
 
-* The original package name 'com.github.gitvoytenko.spring-in-action.toco-cloud' is invalid and this project uses 'com.github.gitvoytenko.springinaction.tococloud' instead.
+* The original package name 'com.github.gitvoytenko.spring-in-action.toco-cloud' is invalid and this project uses 'com.github.gitvoytenko.springinaction.tacocloud' instead.
 
 # Getting Started
